requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
cssselect==1.2.0
cssutils==2.9.0
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.3
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
Pillow==10.1.0
premailer==3.10.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
yagmail==0.15.293